+++ CHANGELOG.md
@@ -6,7 +6,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - ${maven.build.timestamp}
 ### Added
+- `language.questlang_whitelist` to restrict which languages players can select via `/questlang`
 ### Changed
+- `language` config option is now a section; the default language moved to `language.default`
 ### Deprecated
 ### Removed
 ### Fixed